06-google-cloud-function-with-real-case-studies/code/Demo3 - GCS & Face Detection with Vision API/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-vision==2.6.1
google-cloud-storage==1.42.2